modal working and styled
